put request in its on controller files
