+++ client/src/Review/HelpfulWidget.jsx
@@ -4,9 +4,8 @@ import axios from 'axios';
 import StarRating from '../Helper-Components/StarRating.jsx';
 import { Container, FlexContainer } from './Styles/ReviewStyles.js';
 import reviewData from '../../../APIExamples/reviews.js';
+import {helpfulClick} from '../Controllers/reviewController.js'; //'../src/Controllers/reviewControllers.js';
 import API_KEY from '../config/config.js';
-
-
 
 const HelpfulWidget = (props) => {
 
@@ -16,22 +15,7 @@ const HelpfulWidget = (props) => {
 
   //update the helpfulness rating with a put request to the reviewID
   const handleClick = (string) => {
-    let config = {
-      method: 'put',
-      url: `https://app-hrsei-api.herokuapp.com/api/fec2/hr-rfe/reviews/${review.review_id}/${string}`,
-      headers: {
-        'Authorization': API_KEY
-      }
-    };
-
-    axios(config)
-      .then((response) => {
-        console.log(`Successfly reported as ${string}, Status Code ${response.status}`);
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-
+    helpfulClick(review.review_id, string);
   };
 
   return <span>Helpful?{'              '}
